Refine Issue #37
We introduce the new tag "Max" to indicate if the length of this reference variable is known or not. If a variable named like:

`val a_blazeLocalMax4096 = brdcst.data`

It means that the length of this variable is not fixed, so the array length must be provided by the user. For this case, the system will use `for (int localAryCpy = 0; localAryCpy < brdcst_javaArrayLength; localAryCpy++)` to copy the array.

On the other hand, for the variable named:

`var a_blazeLocal4096 = brdcst.data`

It means that the length of the variable is fixed and doesn't have to be sent in runtime. For this case, the system will use `for (int localAryCpy = 0; localAryCpy < 4096; localAryCpy++)` to copy the array.